fix(redis): Raise Base Exception error
