+++ ddtrace/contrib/asyncpg/__init__.py
@@ -50,8 +50,11 @@ required_modules = ["asyncpg"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.asyncpg.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        from ..internal.asyncpg.patch import get_version
+        from ..internal.asyncpg.patch import patch
+        from ..internal.asyncpg.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
