+++ ddtrace/contrib/aws_lambda/__init__.py
@@ -38,9 +38,10 @@ Configuration
 For additional configuration refer to
 `Instrumenting Python Serverless Applications by Datadog <https://docs.datadoghq.com/serverless/installation/python>`_.
 """
-from .patch import get_version
-from .patch import patch
-from .patch import unpatch
+
+from ddtrace.contrib.internal.aws_lambda.patch import get_version
+from ddtrace.contrib.internal.aws_lambda.patch import patch
+from ddtrace.contrib.internal.aws_lambda.patch import unpatch
 
 
 __all__ = ["patch", "unpatch", "get_version"]
