+++ ddtrace/contrib/boto/__init__.py
@@ -35,7 +35,10 @@ required_modules = ["boto.connection"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
+        # Required to allow users to import from `ddtrace.contrib.boto.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        from ..internal.boto.patch import get_version
+        from ..internal.boto.patch import patch
 
         __all__ = ["patch", "get_version"]
