+++ ddtrace/contrib/botocore/__init__.py
@@ -118,8 +118,11 @@ required_modules = ["botocore.client"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import patch_submodules
+        # Required to allow users to import from `ddtrace.contrib.botocore.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        from ..internal.botocore.patch import get_version
+        from ..internal.botocore.patch import patch
+        from ..internal.botocore.patch import patch_submodules
 
         __all__ = ["patch", "patch_submodules", "get_version"]
