+++ tests/contrib/botocore/test_bedrock.py
@@ -5,8 +5,8 @@ import mock
 import pytest
 
 from ddtrace import Pin
-from ddtrace.contrib.botocore.patch import patch
-from ddtrace.contrib.botocore.patch import unpatch
+from ddtrace.contrib.internal.botocore.patch import patch
+from ddtrace.contrib.internal.botocore.patch import unpatch
 from tests.contrib.botocore.bedrock_utils import _MODELS
 from tests.contrib.botocore.bedrock_utils import _REQUEST_BODIES
 from tests.contrib.botocore.bedrock_utils import get_request_vcr
@@ -306,7 +306,7 @@ def test_read_error(bedrock_client, request_vcr):
     with request_vcr.use_cassette("meta_invoke.yaml"):
         response = bedrock_client.invoke_model(body=body, modelId=model)
         with mock.patch(
-            "ddtrace.contrib.botocore.services.bedrock._extract_text_and_response_reason"
+            "ddtrace.contrib.internal.botocore.services.bedrock._extract_text_and_response_reason"
         ) as mock_extract_response:
             mock_extract_response.side_effect = Exception("test")
             with pytest.raises(Exception):
@@ -319,7 +319,7 @@ def test_read_stream_error(bedrock_client, request_vcr):
     with request_vcr.use_cassette("meta_invoke_stream.yaml"):
         response = bedrock_client.invoke_model_with_response_stream(body=body, modelId=model)
         with mock.patch(
-            "ddtrace.contrib.botocore.services.bedrock._extract_streamed_response"
+            "ddtrace.contrib.internal.botocore.services.bedrock._extract_streamed_response"
         ) as mock_extract_response:
             mock_extract_response.side_effect = Exception("test")
             with pytest.raises(Exception):
@@ -333,7 +333,7 @@ def test_readlines_error(bedrock_client, request_vcr):
     with request_vcr.use_cassette("meta_invoke.yaml"):
         response = bedrock_client.invoke_model(body=body, modelId=model)
         with mock.patch(
-            "ddtrace.contrib.botocore.services.bedrock._extract_text_and_response_reason"
+            "ddtrace.contrib.internal.botocore.services.bedrock._extract_text_and_response_reason"
         ) as mock_extract_response:
             mock_extract_response.side_effect = Exception("test")
             with pytest.raises(Exception):
